requirements.txt

azure-ai-documentintelligence==1.0.0b2
azure-core==1.30.1
black==23.12.1
openai==1.6.1
opencv-python==4.5.4.60
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
python-docx==1.1.0
streamlit==1.29.0

